test(db): move sync tests next to module + hoist IndexedBook builders
